requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
# implicit>=0.7.0  # Optional - requires C++ compiler

# Database & Storage
redis>=4.5.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
asyncpg>=0.29.0

# MLOps & Experiment Tracking
mlflow>=2.8.0

# API & Web
requests>=2.31.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Visualization & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0

# Development
pytest>=7.4.0

# Additional ML libraries
lightgbm>=4.0.0